+++ lib/base-compiler.js
@@ -89,7 +89,8 @@ class BaseCompiler {
             timeoutMs: this.env.ceProps("compileTimeoutMs", 7500),
             maxErrorOutput: this.env.ceProps("max-error-output", 5000),
             env: this.env.getEnv(this.compiler.needsMulti),
-            wrapper: this.compilerProps("compiler-wrapper")
+            wrapper: this.compilerProps("compiler-wrapper"),
+            wrapperOptions: this.compilerProps("compiler-wrapper-options")
         };
     }
