+++ lib/cl-support.js
@@ -27,7 +27,8 @@
 const fs = require('fs-extra'),
     path = require('path'),
     _ = require('underscore-node'),
-    utils = require('./utils');
+    utils = require('./utils'),
+    Demangler = require("./demangler").Demangler;
 
 function OverwriteAsmLines(result, demangleOutput) {
     const lines = utils.splitLines(demangleOutput);
@@ -62,6 +63,10 @@ function RunCLDemangler(compiler, result) {
         return compiler.exec(demangler, [], {input: asmFileContent})
             .then(demangleResult => OverwriteAsmLines(result, demangleResult.stdout));
     }
+}
+
+class CLDemangler extends Demangler {
+
 }
 
 exports.RunCLDemangler = RunCLDemangler;
