+++ lib/demangler.js
@@ -30,15 +30,6 @@ const
     exec = require('../lib/exec'),
     logger = require('../lib/logger').logger;
 
-const LabelMetadata = [
-    {'ident': 'C1', 'description': 'complete object constructor'},
-    {'ident': 'C2', 'description': 'base object constructor'},
-    {'ident': 'C3', 'description': 'complete object allocating constructor'},
-    {'ident': 'D0', 'description': 'deleting destructor'},
-    {'ident': 'D1', 'description': 'complete object destructor'},
-    {'ident': 'D2', 'description': 'base object destructor'}
-];
-
 class Demangler {
     /**
      * 
@@ -139,6 +130,14 @@ class Demangler {
         return this.result;
     }
 
+    ExecDemangler() {
+        return exec.execute(
+            this.demanglerExe,
+            [],
+            options
+        );
+    }
+
     Process(result, execOptions) {
         let options = execOptions || {};
         this.result = result;
@@ -150,11 +149,7 @@ class Demangler {
 
         options.input = this.GetInput();
 
-        return exec.execute(
-            this.demanglerExe,
-            [],
-            options
-        ).then((output) => this.ProcessOutput(output));
+        return ExecDemangler().then((output) => this.ProcessOutput(output));
     }
 }
