+++ lib/exec.js
@@ -36,7 +36,9 @@ function execute(command, args, options) {
 
     if (options.wrapper) {
         args = args.slice(0); // prevent mutating the caller's arguments
+        if (command.startsWith('./')) command = path.join(process.cwd(), command);
         args.unshift(command);
+        if (options.wrapperOptions) args.unshift(options.wrapperOptions);
         command = options.wrapper;
 
         if (command.startsWith('./')) command = path.join(process.cwd(), command);
