+++ lib/pascal-support.js
@@ -24,8 +24,9 @@
 "use strict";
 
 const SymbolStore = require("./symbol-store").SymbolStore;
+const Demangler = require("./demangler").Demangler;
 
-class PascalDemangler {
+class PascalDemangler extends Demangler {
     constructor() {
         this.symbolStore = new SymbolStore();
         this.fixedsymbols = {};
